#119 Rename module (core.aspandas → core.api)
